Store background process pids in linked list
This creates a linked list when the program starts in which to store
background process IDs. The linked list is passed to the various
processing functions and the PIDs are added.

Just need to add waitpid calls to check/report the status of these processes
and remove them when finished or killed.